Modify constants

+++ containers.h
@@ -501,7 +501,7 @@ unordered_set unordered_set_init(size_t key_size,
 
 /* Utility */
 int unordered_set_rehash(unordered_set me);
-int unordered_set_size(unordered_set me);
+size_t unordered_set_size(unordered_set me);
 int unordered_set_is_empty(unordered_set me);
 
 /* Accessing */
